+++ lib/shrine/plugins/sequel.rb
@@ -62,7 +62,7 @@ class Shrine
           # reload the attacher on record reload
           define_method :_refresh do |*args|
             result = super(*args)
-            send(:"#{name}_attacher").reload
+            send(:"#{name}_attacher").reload if instance_variable_defined?(:"@#{name}_attacher")
             result
           end
           private :_refresh